+++ src/ship.cpp
@@ -4,12 +4,12 @@
 
 using namespace std;
 
-Ship::Ship(vector<glm::vec3>& map): _map(map) {
+Ship::Ship(vector<glm::dvec3>& map): _map(map) {
     load("spaceship");
 	_shaderNb = 1;
 	_gameOver = false;
-    translate(glm::vec3(0., 0., SHIP_HEIGHT));
-    _roadPosition = glm::vec3(0., 0., SHIP_HEIGHT);
+    translate(glm::dvec3(0., 0., SHIP_HEIGHT));
+    _roadPosition = glm::dvec3(0., 0., SHIP_HEIGHT);
     _orientation = 0.;
     _speed.x = 0.;
     _abSpeed = 0;
@@ -26,48 +26,48 @@ Ship::Ship(vector<glm::vec3>& map): _map(map) {
 Ship::~Ship() {
     }
 
-glm::vec3 Ship::getVertical() {
-    return glm::vec3(_modelMatrix[2]);
+glm::dvec3 Ship::getVertical() {
+    return glm::dvec3(_modelMatrix[2]);
     }
 
-void Ship::turn(float angle) { //Angle in degrees
-    rotate((float)glm::radians(angle), glm::vec3(0., 0., 1.));
+void Ship::turn(double angle) { //Angle in degrees
+    rotate((double)glm::radians(angle), glm::dvec3(0., 0., 1.));
     _orientation += angle;
     }
 
-void Ship::thrust(float x) {
-    _speed += glm::vec2(x * cos(glm::radians(_orientation)), x * sin(glm::radians(_orientation)));
+void Ship::thrust(double x) {
+    _speed += glm::dvec2(x * cos(glm::radians(_orientation)), x * sin(glm::radians(_orientation)));
     }
 
-float Ship::getSpeed() {
+double Ship::getSpeed() {
     return _abSpeed;
     }
 
-void Ship::setFriction(float pitchGoal){
+void Ship::setFriction(double pitchGoal){
         _frictionCoef = (ACCELERATION / std::pow(SPEED_REFERENCE * pitchGoal, DECCELERATION_FACTOR));
 }
 
 void Ship::manage() {
 	if(_gameOver){
-		translate(glm::vec3(0., 0., -1));
+		translate(glm::dvec3(0., 0., -1));
 		_roadPosition.z -= 1.;
 		return;
 	}
-    translate(glm::vec3(0., 0., -_roadPosition.z));
+    translate(glm::dvec3(0., 0., -_roadPosition.z));
     _counter += SHIP_ROUTINE_STEP;
 
     if (_counter >= 360)
         _counter -= 360;
 
     _roadPosition.z = SHIP_HEIGHT + SHIP_ROUTINE_AMPLITUDE * sin(glm::radians(_counter));
-    translate(glm::vec3(0., 0., _roadPosition.z));
+    translate(glm::dvec3(0., 0., _roadPosition.z));
 
-    float hypo = sqrt(_speed.x * _speed.x + _speed.y * _speed.y);
+    double hypo = sqrt(_speed.x * _speed.x + _speed.y * _speed.y);
     if (hypo != 0.) {
         _inertiaAngle = glm::degrees(asin(_speed.y / hypo));
         move(hypo);
-        _speed = glm::vec2(hypo * cos(glm::radians(_inertiaAngle)), hypo * sin(glm::radians(_inertiaAngle)));
-        glm::vec2 dec(_speed);
+        _speed = glm::dvec2(hypo * cos(glm::radians(_inertiaAngle)), hypo * sin(glm::radians(_inertiaAngle)));
+        glm::dvec2 dec(_speed);
 		dec *= _frictionCoef * std::pow(hypo, DECCELERATION_FACTOR - 1);
         _speed -= dec;
         }
@@ -81,9 +81,9 @@ void Ship::manage() {
 		_uniform[0] = 1.;
     }
 
-void Ship::move(float x) { //x MUST be positive
-    float deltaX = x * cos(glm::radians(_inertiaAngle));
-    float diff, adv;
+void Ship::move(double x) { //x MUST be positive
+    double deltaX = x * cos(glm::radians(_inertiaAngle));
+    double diff, adv;
     _abSpeed = 0; //We are going to evaluate the X-distance the ship is making during this move
 
     //PHYSICS MADAFAKA
@@ -95,10 +95,10 @@ void Ship::move(float x) { //x MUST be positive
 
         //Moving to the next segment
         //We are not moving forward but towards inertia point
-        rotate(glm::radians(_inertiaAngle - _orientation), glm::vec3(0., 0., 1.));
-        translate(glm::vec3(adv, 0., 0.));
+        rotate(glm::radians(_inertiaAngle - _orientation), glm::dvec3(0., 0., 1.));
+        translate(glm::dvec3(adv, 0., 0.));
         //Restoring initial orientation
-        rotate(glm::radians(_inertiaAngle - _orientation), glm::vec3(0., 0., -1.));
+        rotate(glm::radians(_inertiaAngle - _orientation), glm::dvec3(0., 0., -1.));
         _roadPosition.y += -adv * sin(glm::radians(_inertiaAngle)); //New Y-position
 
 
@@ -114,24 +114,24 @@ void Ship::move(float x) { //x MUST be positive
         _inertiaAngle -= glm::degrees(_map[_curSegment][1]); //Opposite orientation is added : new road-relative orientation
 
         //Bringing the ship at the frontier, facing the new segment
-        translate(glm::vec3(0., 0., -_roadPosition.z));
-        rotate((float)glm::radians(_orientation), glm::vec3(0., 0., -1.)); //
+        translate(glm::dvec3(0., 0., -_roadPosition.z));
+        rotate((double)glm::radians(_orientation), glm::dvec3(0., 0., -1.)); //
 
         //Applying new Y-rotation, ship is now in the new segment's plane
-        rotate((float)_map[_curSegment][2], glm::vec3(0., 1., 0.));
+        rotate((double)_map[_curSegment][2], glm::dvec3(0., 1., 0.));
 
         //Restoring road-relative position and orientation
-        rotate((float)glm::radians(_orientation), glm::vec3(0., 0., 1.)); //
-        translate(glm::vec3(0., 0., _roadPosition.z));
+        rotate((double)glm::radians(_orientation), glm::dvec3(0., 0., 1.)); //
+        translate(glm::dvec3(0., 0., _roadPosition.z));
 
         //Calculating remaining distance to move
         x -= adv; //Relative
         deltaX = x * cos(glm::radians(_inertiaAngle)); //Absolute
         }
 
-    rotate(glm::radians(_inertiaAngle - _orientation), glm::vec3(0., 0., 1.));
-    translate(glm::vec3(x, 0., 0.)); //Movement of remaining distance
-    rotate(glm::radians(_inertiaAngle - _orientation), glm::vec3(0., 0., -1.));
+    rotate(glm::radians(_inertiaAngle - _orientation), glm::dvec3(0., 0., 1.));
+    translate(glm::dvec3(x, 0., 0.)); //Movement of remaining distance
+    rotate(glm::radians(_inertiaAngle - _orientation), glm::dvec3(0., 0., -1.));
 
     _roadPosition.x += deltaX; //Adding remaining absolute distance
     _abSpeed += deltaX; //Idem
@@ -139,8 +139,8 @@ void Ship::move(float x) { //x MUST be positive
     _absPos += _abSpeed; //Updating general X-position
     }
 
-glm::vec3 Ship::getAbsPos() {
-    return glm::vec3(_absPos, _roadPosition.y, _roadPosition.z);
+glm::dvec3 Ship::getAbsPos() {
+    return glm::dvec3(_absPos, _roadPosition.y, _roadPosition.z);
     }
 
 void Ship::putOnRoad() {
